requirements.txt
Flask==2.2.2
google_api_python_client==2.102.0
google_auth_oauthlib==1.1.0
protobuf==3.19.1
Requests==2.31.0
waitress==2.1.2
